feat(subscriber): implement recording format reading and writing
Added the very first recording format and functionality to read and
write the format. Writing from the `RfrLayer` has been enabled, but it's
very basic and re-writes the recording file on every span or event call
that is received by the layer.

The recording file is stored as a binary file using the
[postcard](https://github.com/jamesmunns/postcard) format.

This functionality has been placed in the `rfr-subscriber` for now, but
likely should be moved to a separate shared crate later on.