feat(viz) add `rfr-viz` crate
This crate will create visualization based on flight recording data. The
crate itself currently only contains hello world code.

The main addition in this commit is a mock up of a visualization of
task and waker instrumentation on a timeline. It is the basis for a
proof of concept visualization.