refac(rfr,subscriber): move rec definitions to `rfr` crate
The definitions of the recording file format (wire format) were
initially placed in the `rfr-subscriber` crate. This doesn't make sense,
as any other crate that wants to read the file would then need to depend
on `rfr-subscriber`.

This change moves teh format definitions and file reading and writing
functions to the `rfr` crate instead. This will provide a common
dependency for any crate that wants to deal with flight recording files,
either writing or reading them.