feat(viz): add very simple visualization from rfr file
Add a very basic visualization of task states. It's essentially the
minimal implementation that shows something based on the reference
visualization HTML.

It currently uses a fixed 100µs per pixel, but in a sort of hacky way.
This will now be the basis for refactoring into a better version and
including spawn and waker annotations.