refac(viz): collect waker and spawn events (#1)
In the initial dummy implementation, only task events (excluding wakes)
were processed and visualized.

Visualizing wake and spawn events is somewhat more complicated, it is
made additionally so if we want to be more robush when collecting the
task events, and include waking actions in those task events.

This commit is the first step towards doing so.

We assign all the events for a given task to that task (thus removing
the fake `Task` event from the list). The remaining event records are
then converted into 3 different event types:
- TaskEvent: this includes new, drop, and poll start and end as well as
  wake and wake by ref events.
- WakeEvent: this includes wake, wake by ref, clone, and drop wake
  events. Additionally, the wake and wake by ref events are divided into
  normal and self-* events (so there are 6 kinds in total)
- SpawnEvent: this is the event where a task is spawned. A task can have
  0 or 1 of these.

The TaskEvents are then walked and a list of TaskSection elements are
created. These task sections can be used directly to construct the task
stte bar in the visualization. The WakeEvent objects and SpawnEvent
object can be used directly to visualize the waker markers and spawn
marker respectively.

This change also adds a hover style to the spawn and waker markers such
that they come in front of other markers (there is often overlap because
the markers are close together) and shows a border around the marker
with the same color as the line and arrows.