refac(rec,subscriber): only write rfr stream file
In the initial implementation of the `rfr-subscriber` layer, the records
created after each tracing function was called were accumulated in a
buffer and then the entire buffer was written out each time.
Additionally, each record was printed to `stdout`.

While useful for debugging, this wasn't very efficient.

This change gets rid of the file stored on the layer entirely and gives
that over to the `rfr::rec::StreamWriter` (previously just `Writer`).
Internally the `StreamWriter` wraps the inner writer in a `BufWriter`.
This means that the code that creates an `RfrLayer` will need to get a
"flusher" from it to flush the `BufWriter` before exiting the process.

Since the `StreamWriter` is dedicated to writing a stream of events, we
cannot know how many there will be before time. So the file format has
been changed to remove the element count (which allows us to treat the
postcard data as a `Vec` with a known length). The code to read a stream
file has also been updated to handle reading records until the end of
the file (since it no longer has a record count to rely on).

This is an imcompatible change to the file format, so it has been
updated from `rfr/0.0.1` to `rfr/0.0.2`.