docs: add initial book
The content of the book so far describes the purposes of RFR, as well as
file formats.

The encoding links to Postcard and the (new) format identifier is
defined.

A new version (`rfr-s/0.0.3`) of the streaming file format is defined.
It is only slightly different from the current format.

The anticipated chunked file format is also described. This has not yet
been implemented at all.

Those types which are shared between both file formats are listed in a
common sub-chapter.