feat(rfr): implement string `FormatIdentifier`
The new RFR Book (which lives in the `docs` directory of this repo)
describes a format identifier which is stored as a string at the
beginning of each RFR file. However, this is not what had previously
been implemented.

This change renames the `FormatVersion` struct to `FormatIdentifier` and
adds implementations to write to and read from a string representation.
This is then used when writing and reading the implemented streaming
format. As such, `FormatIdentifier` and `FormatVariant` no longer derive
the `Serialize` and `Deserialize` traits, as neither should be used for
(de)serialization directly.

The main advantage of the string representation is that it can be
verified by inspecting the first few characters of an RFR file
"manually" in a text editor or similar.