feat(rfr,subscriber): add initial chunked implementation (#5)
An important feature of RFR is writing recordings as a series of self
contained chunks, which can be read individually and without knowledge
of other chunks.

This commit adds the initial implementation to write chunks to disk,
including both the data structures in the `rfr` crate as well as a new
`tracing-subscriber` Layer in the `rfr-subscriber` crate which will
write chunks from the tracing instrumentation in tokio.

A new example has been added to the `rfr-subscriber` which runs the
"spawn" example, but with the `RfrChunkedLayer`.

Currently, the chunks are not written on a schedule, as is planned.
Instead, all the thread chunks are stored until the layer's flusher has
flush called on it. This is only an intermediate state. Additionally,
there is no support for reading a chunked flight recording yet.