feat(rfr,viz): read and visualize chunked recordings (#6)
The `rfr` crate now has a `Recording` struct that is able to lazy load a
recording made up of multiple chunks. It then exposes simple versions of
the data structures in the chunks which can be used to read out the
data.

As part of this, the `rfr-viz` crate has been modified to be generic
over the format of recording, streaming or chunked. This was done by
converting both formats to a standard representation of "task + event"
objects, which is not that far off what the streaming format was already
creating.

The `rfr-viz` binary will now pick the recording format to decode based
on whether the parameter that it is given is a plain file (streaming) or
a directory (chunked). This is a little hacky, but it will work for the
time being.

This binary also outputs a bunch of intermediate debug output to stdout.
This is left for the time being as an aid to future work (although it
should really be removed at some point).

A new ping-pong example using the chunked tracing-subscriber layer has
been added to test.