refac(rfr): rename ThreadChunk to SeqChunk
To make the chunked recording format more general, the `ThreadChunk`
struct and associated `ThreadChunkBuffer` have been renamed to
`SeqChunk` and `SeqChunkBuffer` respectively. Here "seq" is short for
sequence and the implementation has been moved into a `sequence`
sub-module.

Part of the reasoning for this is to reenforce the point that a sequence
chunk should have all the events ordered in a single sequence. Mostly,
this is done by taking events from a single thread, which is a sequenced
execution.

Additionally, sequence chunks have been given a sequence identifier
(`SeqId`). This will allow sequence chunks from different chunks to be
matched together into a single sequence.