chore: add GitHub CI workflows (#7)
Add a CI workflow configuration to build, test, and lint the rfr project
on GitHub.

The build jobs test Rust stable on macOS, Windows, and (Ubuntu) Linux as
well as nightly, and Rust 1.76.0 on Linux. The MSRV was chosen based on
what the code already written supports (and may change before any
release).

A separate job has been added which runs 4 examples (spawn and ping-pong
each with streamed and chunked recordings) and then gives the output to
`rfr-viz`. This isn't checking correctness, just for panics.

The CI file is borrowed from the tokio-rs/console project and modified
to suit our needs. A small change to make Clippy happy is also included
and another to use an imported instead of absolute module path for
`std::ptr` (which is related to the thing that didn't allow us to take
an MSRV of 1.74.0).