refac(rfr): clean up chunk headers (#10)
The chunk and sequence chunk headers were a bit of a mess. The naming
wasn't consistent and some information was stored either in the wrong
format, was missing, or wasn't needed.

This cleans the headers up.

A chunk now has the following information:
- interval: the time period the chunk is responsible for
- earliest timestamp: first timestamp recorded within the contained
  sequence chunks
- latest timestamp: latest timestmap recorded within the contained
  sequence chunks

A sequence chunk contains the following information:
- sequence ID: an identifier for this sequence across chunks (it's a
  sort of RFR-specific thread ID in the current implementation).
- earliest timestamp: first timestamp recorded in this sequence
- latest timestamp: last timestamp recorded in this sequence

An interval is a base time as well as a (chunk relative) start and end
time. It should always be a multiple of a second, or divide 1 second
without remainder.

The sequence chunk buffer stores the interval of the chunk it belongs
to, to keep a thread local copy of this information. However, a sequence
chunk, as stored in a recording, doesn't contain this information as it
is the same for all sequence chunks in the same chunk.