fix(viz): fix reading from multi-threaded chunked recordings
There was an error in `rfr-viz` when reading from chunked recordings
of a multi-threaded tokio runtime. This was due to overwriting events
from one sequence with those from another, instead of appending (and
then sorting) them into a single list.

This change fixes the implementation error in `rfr-viz` and switches the
two `rfr-subscriber` examples which emit chunked recordings to use
multi-threaded tokio runtimes with 2 worker threads.

The `rfr-viz` code was also modified to reduce the `stdout` output a bit
(but there's still quite a lot).