feat(subscriber): write chunks as they are completed (#13)
The chunked recording format splits the recording into separate files -
each one representing one chunk of the recording, which is a specific
time interval.

The chunked tracing layer collected the chunks, but they were only
written at the end of the execution. Until that point, all chunks were
kept in memory - which loses the whole benefit of the chunked format.

This change adds support to the `ChunkedWriter` to write only completed
chunks out. The method then provides a duration until the next chunk
would be ready. The chunked layer uses this new interface, spawning a
dedicated thread to write out completed chunks and then sleeping until
the next chunks will be ready.

Additionally, a new example has been added which creates a Tokio Barrier
and synchronizes on it.