docs(rfr): describe format for storing spans and events (#14)
The current version of the RFR file format records tasks as well as task
lifecycle events and waker events. This is already useful to describing
async applications, but is quite narrow.

One often requested feature of Tokio Console is to show spans and events
that occur within the context of tasks. This would allow a developer to
further instrument their own code and see that instrumentation overlaid
with the information about how the async runtime is operating.

This change to the RFR file format includes the necessary definitions to
record generic spans and events. While it stays quite close to the model
that tracing uses, some effort has been made to generalize the
definitioin.

Since in the current implementation, spans and tasks share an ID, this
has been included as the `InstrumentationId`, which is what will link
record data to objects. This will mean that a streaming implementation
no longer needs to keep an instrumentation ID to task ID map for all
active tasks.

Callsites will also be stored separately (since they are often shared
between many objects). In the chunked format, this storage will be
central for the entire recording, as the storage size is expected to be
minimal and it reduces the per chunk overhead, both for processing and
storage.

As we will now have a type of what was previously called `Event` which
represents a tracing event, some renaming was necessary. In the chunked
file format an `EventRecord` is now called a `Record` and and `Event` is
now called `RecordData`. In the streaming file format, an `Event` is now
`RecordData` so the chunked and streaming formats are more closesly
aligned.

Additionally, separate format identifiers have been specified for the
chunked file format meta file as well as a new file for storing
callsites (centrally for an entire chunked recording).

No code changes are included in this update.