rfr: impl De/Serialize for `FormatIdentifier` (#15)
The `FormatIdentifier` struct is used as a header in streaming and
chunked recording formats. The format identifier is defined as
serializing to a string format, which makes it easy for a human to read
it out when looking at the binary file in a hex editor or just a text
editor.

However, the `FormatIdentifier` struct didn't implement `Serialize` or
`Deserialize` (from `serde`), because with this string format, the
derive macros don't give us what we want. Instead we wrote it out to a
string format and then serialized or deserialized to a string and then
convert.

This change "manually" implements `Serialize` and `Deserialize` to write
it out to or read it in from the string representation directly. The
code in `rfr` has been updated to use calls to `postcard` directly.