feat: add format identifiers to chunked meta file
In the recording format changes described in #14, a list of format
identifiers was added to the chunked recording meta file. Additionally,
separate format identifiers were specified for each file type, such that
the meta file gets its own format identifier.

The list of format identifiers contains all the (other) format
identifiers used in the recording, allowing software that is attempting
to read the chunked recording to check compatibility based on the meta
file alone.

This change updates reading and writing of the meta file to match the
new format. Additionally, the structs and logic for the meta file have
been moved to a dedicated module, as we slowly clean up the mess of code
in the `chunked` module.