feat(rfr): add callsites file
The general callsite information for a span, event, or other
instrumentation objection doesn't change during the execution of an
application. Additionally, the number of callsites is bounded (these are
effectively instrumented points in the code). As such, it makes sense to
store that information centrally for a chunked recording.

The object data which needs to be stored in each sequence chunk can then
be reduced.

This change is the first step towards this format (which was documented
in #14 and #17). The `Callsite` struct has been defined, as well as the
structure of the central `callsites.rfr` file together with the
necessary functionality to write this file incrementally (as callsites
are recorded for the first time).

This has been connected to the `ChunkedWriter` and will cause an empty
callsites file to be created (containing just the format identifier),
but not callsite data will be recorded or written out, as the common
`rfr` objects need to be modified to use the `Callsite` first.