feat(rfr): support generic spans and events in file format
WIP: This is a work in progress, based on changes to the design in #14
and #17.

Implement the changes designed in #14 and #17 which bring support for
generic spans and events into the recording file format. This is in
addition to the task spans and waker events already supported.

In this change, only the file structure changes are made, together with
all the necessary changes in the `rfr-subscriber` and `rfr-viz` crates
to support the modification to the format.

Generic spans and events are not yet actually recorded.