feat(rfr): support generic spans and events in file format
Implement the changes designed in #14 and #17 which bring support for
generic spans and events into the recording file format. This is in
addition to the task spans and waker events already supported.

In this change, only the file structure changes are made, together with
all the necessary changes in the `rfr-subscriber` and `rfr-viz` crates
to support the modification to the format.

Generic spans and events are not yet actually recorded.

That said, it's a big refactor, especially since what we now call a
`Record` was previously called an `Event` (renamed because we have
events which are like tracing events now) and what is now called
`RecordData` was previously called `EventRecord`.

There is also a small change in the docs for the streaming variant of
`RecordData` is made (using structs instead of tuples for data) to align
it with the chunked variant.