refac(viz): rename events to records
When the changes in the `rfr` crate to rename `EventRecord` to `Record`,
much of the naming in the `rfr-viz` crate no longer made sense.

This change renames `SpawnEvent`, `WakerEvent`, and `TaskEvents` to
`SpawnRecord`, `WakerRecord`, and `TaskRecords` respectively along with
the related `*Kind` enums.

The change should have no functional impact.