chore: use justfile for running examples (#22)
There are a number of examples in the `rfr-subscriber` crate which show
some basic usage of rfr. As a sanity check (because they are by no means
exhaustive), all these examples are run and then their output is passed
to `rfr-viz` to be read.

In order to make running those examples locally easier, they have been
added to a `justfile` to be used with the just command runner.

More details on just: https://github.com/casey/just

The CI examples job has been modified to use the `justfile`.