add spawn markers and lines
This change adds the spawn marker at the beginning of a task status bar.
If the task was spawned by another task (that we know of), then a line
is drawn from that other task to the spawned task.

Spawn markers and lines are drawn on top of all tasks, but the lines may
be drawn over some other task's spawn marker, which we may not want. To
avoid that, we'd have to draw in 2 passes, one pass for the lines and
another for the markers.