feat(viz): add the `rfr-viz` GUI app (#23)
One of the goals of RFR is to have rich visualization of task details
available. Until now the `rfr-viz` app has generated HTML files which
served as a proof of concept for this visualization, however they lacked
interactivity.

This change adds a very basic GUI app as an alternative way to use
`rfr-viz`. It takes a flight recording and then displays it in an [egui]
app. The current functionality matches the generated HTML (which can
still be produced), with no additional features.

The current implementation should be considered a proof of concept, the
code is pretty rough.

[egui]: https://github.com/emilk/egui