feat(subscriber): add thousand tasks example (#24)
Most of the instrumented examples (in the `rfr-subscriber` crate) have a
relatively small number of tasks. It would be useful to have a standard
example with a few more.

The `thousand-tasks` example has just that, 1000 spawned tasks (plus the
usual compliment of a block_on task and blocking tasks for blocking
workers).

Given the lack of optimization (or any sort of performance consideration
at all) it runs surprisingly well in the RFR Viz UI.