fix(viz): display final state of task (#26)
When a task is not dropped by the end of the recording, we were not
displaying a section for its final state. This was especially obvious in
Tokio's blocking workers, which show up as blocking tasks that become
active and then never change, however it will also affect any partial
recording, where it is likely that some tasks would still exist at the
end of the available recordering.

To solve this problem, the `TaskRow` struct had a new `final_state`
field, which optionally holds the last known state of the task. In the
Viz UI, we then create a section which extends until the end of
recording time with this state.

The Viz HTML generator was also updated to extend the last state as a
section to the end of the visualization.