fix(rfr): make `Event` fields public
The fields of all the structs and enums that represent a recording
(streaming or chunked) should be public. This doesn't include the helper
reader and writer structs, just those that implement Serialize and
Deserialize.

This wasn't the case for the `Event` struct, where none of the fields
were marked `pub`.

This change fixes this and marks all fields `pub`.