chore: split examples from HTML gen in justfile (#28)
With the new Viz UI, we often want to only run examples with `just`, and
not run `rfr-viz` to generate HTML (since we'll use the Viz UI instead).

This change splits all the targets in the `justfile` so that each
example can be run separately from the Viz HTML generation. The previous
targets that run the example and then generate the HTML directly still
exist, but now depend on the split targets to achieve that.