feat(rfr): add method to wait for a chunk to be written (#29)
The `ChunkedWriter::write_all_chunks` method produces inconsistent
sequence chunks. They may contain records which occur after the sequence
chunk's `latest_timestamp`. This occurs because the latest timestamp is
recorded before all the individual events are written to disk, and new
records may come in during this time interval.

To solve this problem, this change adds a new method on the
`ChunkedWriter` to wait until the "next" chunk gets written. This is
slightly more complicated as we actually wait until the chunk which will
contain a record from the current time (when called) will be written.
The purpose is that we can guarantee that anything recorded prior to
`ChunkWriter::wait_for_write_timeout` being called will have been
written out before this method returns.

The `write_all_chunks` method should probably be removed entirely, but
it hasn't been yet.

The chunked tracing Layer in `rfr-subscriber` has been updated to
provide a Flusher that exposes this functionality.